Added action button inside observation page
✨  Update List ✨
✅  Django: Added update_observation_by_product_id
✅  Xamarin: Changed observations page delete button: now it is an action button
✅  Xamarin: Added UpdateOrDelete function
✅  Xamarin: Added Modify Observation class
✅  Xamarin: Removed not used comments

+++ Xamarin/XamarinFrontEnd/UserObservations.xaml.cs
@@ -53,42 +53,76 @@ namespace XamarinFrontEnd
 
         }
 
-        private async void Delete(object sender, EventArgs e)
+        private async void UpdateOrDelete(object sender, EventArgs e)
         {
             Button button = (Button)sender;
             string observation_product = (string)button.CommandParameter;
-            HttpResponseMessage response = await ObservationRequest.DeleteObservation(observation_product);
 
-            if (response.IsSuccessStatusCode)
+            string action = await DisplayActionSheet("What would you like to do?", "Cancel", null, "Update Observation", "Delete Observation");
+
+            if (action == "Delete Observation")
             {
-                try
+                HttpResponseMessage response = await ObservationRequest.DeleteObservation(observation_product);
+                if (response.IsSuccessStatusCode)
                 {
-                    RequestObservation obs_to_remove = CompleteObservations.Find(RequestObservation => RequestObservation.Product.Id == observation_product);
-                    CompleteObservations.Remove(obs_to_remove);
-                    Console.WriteLine(CompleteObservations);
-                    MyCollectionView.ItemsSource = null;
-                    MyCollectionView.ItemsSource = CompleteObservations;
-                    await DisplayAlert("Success!", "Observation removed", "OK");
+                    try
+                    {
+                        RequestObservation obs_to_remove = CompleteObservations.Find(RequestObservation => RequestObservation.Product.Id == observation_product);
+                        CompleteObservations.Remove(obs_to_remove);
+                        Console.WriteLine(CompleteObservations);
+                        MyCollectionView.ItemsSource = null;
+                        MyCollectionView.ItemsSource = CompleteObservations;
+                        await DisplayAlert("Success!", "Observation removed", "OK");
+                    }
+                    catch (Exception ex)
+                    {
+                        await DisplayAlert("Error!", "Something went wrong", "OK");
+                    }
                 }
-                catch (Exception ex)
+                else
                 {
-                    await DisplayAlert("Error!", "Something went wrong", "OK");
+                    if (response.StatusCode == System.Net.HttpStatusCode.BadGateway)
+                    {
+                        await DisplayAlert("Attention!!!", "No connection with the server", "OK");
+
+                    }
+                    if (response.StatusCode == System.Net.HttpStatusCode.BadRequest)
+                    {
+                        await DisplayAlert("Try Again!", "Invalid request", "OK");
+                    }
                 }
             }
-            else
-            {
-                if (response.StatusCode == System.Net.HttpStatusCode.BadGateway)
-                {
-                    await DisplayAlert("Attention!!!", "No connection with the server", "OK");
 
-                }
-                if (response.StatusCode == System.Net.HttpStatusCode.BadRequest)
+            else if (action == "Update Observation")
+            {
+
+                RequestObservation obs_to_update = CompleteObservations.Find(RequestObservation => RequestObservation.Product.Id == observation_product);
+                string threshold_price = await DisplayPromptAsync("What's your desired price?", "Insert a threshold price", keyboard: Keyboard.Numeric);
+                ModifyObservation update_observation = new ModifyObservation(obs_to_update.Email, obs_to_update.Product.Id, threshold_price);
+                string json = JsonConvert.SerializeObject(update_observation);
+                HttpResponseMessage response = await ObservationRequest.UpdateObservation(json, observation_product);
+                if (response.IsSuccessStatusCode)
                 {
-                    await DisplayAlert("Try Again!", "Invalid request", "OK");
+                    int index = CompleteObservations.FindIndex(RequestObservation => RequestObservation.Product.Id == observation_product);
+                    CompleteObservations[index].Threshold_price = string.Format("{0:f2}", threshold_price);
+                    MyCollectionView.ItemsSource = null;
+                    MyCollectionView.ItemsSource = CompleteObservations;
+                    await DisplayAlert("Success!", "Observation updated", "OK");
+                }
+                else
+                {
+                    if (response.StatusCode == System.Net.HttpStatusCode.BadGateway)
+                    {
+                        await DisplayAlert("Attention!!!", "No connection with the server", "OK");
+
+                    }
+                    if (response.StatusCode == System.Net.HttpStatusCode.BadRequest)
+                    {
+                        await DisplayAlert("Try Again!", "Invalid request", "OK");
+                    }
                 }
             }
         }
-
 
     }
